Change to -> Rooms and children, also to Menu
I put in descriptions for the Rooms, cleaned up the Room a little, and
Made Menu a more engauging text to read. As well as a little story. I
need Items and Enemies to be done before i finish my classes. So I will
move onto Enemies and how they are created.