Hallway / Room / EnemyFactory
-Hallway Has a test method if you run it as is then you can see that the
hallway prints out what is inside of it. It is working just as I hoped

-EnemyRandomFactory is a class for the room's to make a random enemy out
of the enemies we have made. This is easier as we can make a great deal
of generic enemies and apply them to all the rooms.

-Room/Subrooms the Armory is the only class that is fully filled out
because i wanted to make sure that i have one room right before copy and
pasting the code to the other rooms. There is 2 Values in the rooms
Max_Items and Max_Enemies these 2 values control the max amount of each
thing you can get in a room. ( its based of a random roll over all )

-Now All I have to do is create the flow of the game, when more items or
more enemies are added to the game it will not effect my code or the
games over all code at all. We do need more enemies and Items, but this
should be that hard.